Consolidate message parsing into textfile module
Remove the messages.py module (inline config messages, disabled) and fold
parse_message/get_display_constants helpers directly into textfile.py.
Also drop the messages entry from config.json.

One module, one mechanism: edit messages.txt to change display content.
